backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gymnasium==0.29.1
stable-baselines3==2.3.2
torch==2.4.1
python-dotenv==1.0.1
requests==2.32.3
pytest==8.3.3
python-multipart==0.0.6